+++ src/display.js
@@ -1,7 +1,11 @@
-/* eslint-disable import/no-cycle */
-import add from './addElement.js';
-import { deleteComplete } from './deleteElements.js';
-import editTask from './editTask.js';
+import Task from './listTask.js';
+import saveList from './saveLocal.js';
+
+const updatePosition = (newList) => {
+  newList.forEach((task, id) => {
+    task.index = id;
+  });
+};
 
 const display = (list) => {
   const title = document.createElement('h4');
@@ -13,7 +17,13 @@ const display = (list) => {
   input.placeholder = 'Add to your list...';
 
   input.addEventListener('keypress', (event) => {
-    add(event, input, list);
+    if (event.key === 'Enter' && input.value !== '') {
+      const newTask = new Task(input.value, false, list.length);
+      list.push(newTask);
+      input.value = '';
+      display(list);
+      saveList(list);
+    }
   });
 
   const listElements = document.createElement('ul');
@@ -28,7 +38,37 @@ const display = (list) => {
     description.innerHTML = element.description;
 
     description.addEventListener('click', () => {
-      editTask(description, element, list);
+      const taskInput = document.createElement('input');
+      taskInput.type = 'text';
+      const delet = document.createElement('input');
+      delet.type = 'button';
+      delet.value = 'delete';
+      delet.addEventListener('click', () => {
+        list.forEach((ele) => {
+          if (ele.index === element.index) {
+            const indice = list.indexOf(ele);
+            list.splice(indice, 1);
+            updatePosition(list);
+            saveList(list);
+            display(list);
+          }
+        });
+      });
+
+      taskInput.value = description.innerHTML;
+      description.parentNode.replaceChild(taskInput, description);
+      taskInput.parentNode.appendChild(delet);
+      taskInput.focus();
+
+      taskInput.addEventListener('keypress', (event) => {
+        if (event.key === 'Enter') {
+          description.innerHTML = taskInput.value;
+          taskInput.parentNode.replaceChild(description, taskInput);
+          element.description = taskInput.value;
+          saveList(list);
+          description.parentNode.removeChild(delet);
+        }
+      });
     });
 
     const checkBox = document.createElement('input');
@@ -51,7 +91,13 @@ const display = (list) => {
   button.type = 'button';
   button.id = 'delete-button';
 
-  button.addEventListener('click', () => deleteComplete(list));
+  button.addEventListener('click', () => {
+    const newList = list.filter((task) => task.completed === false);
+    list = newList;
+    updatePosition(list);
+    saveList(list);
+    display(list);
+  });
 
   listElements.appendChild(button);
